Finished Removal of Persisted Scene. - Cut down on functionality in GameSceneController - it no longer does additive loading and is no longer a monobehavior - Got rid of Launcher and DebugRoomLauncher - Added folder Resources/Singletons.  If you want a resource to be immediately available when the game starts, put it in here - Changed the editor menu from "Play from Persisted Scene" to "Play from Title Scene" - Moved the IngredientDrawer out of LootTableDef and into the Editor file (so the game can build as an executable)
