+++ Assets/Editor/IngredientDrawer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bebf7cc7c900c2048b88bda546cfbf91
+guid: 0ee6af1b9a7d3324ea06a96386c4393a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
