--- Assets/Engine/Persisted.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 30490a2c466cc9443a30eb1f1363c2fe
-folderAsset: yes
-timeCreated: 1524150005
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
