--- Assets/Engine/Persisted/Launcher.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 36e18fe3babfa824ba2e4040f6160aca
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
