--- Assets/Engine/Persisted/PersistedScene.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 501722caa83a5b24cb595cdcabf7256f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
