+++ Assets/Resources.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 9d18df8fb3ee4574499cd119d5962f50
-timeCreated: 1524150005
-licenseType: Free
+guid: 18d80e09ebf99fd499bbfc7463c29ead
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
