+++ Assets/Resources/Singletons.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6d3ffb4d52396e44480eeddca65a7385
+guid: a7369c1aa1ce8544baa0403c7b24145b
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
