+++ Assets/Resources/Singletons/GameSceneController.asset
@@ -8,7 +8,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 845e1d5e091e8a64b8e951e3f40a62c8, type: 3}
-  m_Name: PersistedScene
+  m_Script: {fileID: 11500000, guid: 9b0739eb985afca418fa67b3cab4148d, type: 3}
+  m_Name: GameSceneController
   m_EditorClassIdentifier: 
-  SceneName: PersistedScene
