+++ Assets/Resources/Singletons/GameSceneController.asset.meta
@@ -1,10 +1,8 @@
 fileFormatVersion: 2
-guid: 076042f798816474197c13cf6dfe7a91
-timeCreated: 1524150008
-licenseType: Free
+guid: 2a6f2a7d74101b9419a81dd93d099aef
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 0
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
