+++ Assets/Resources/Singletons/GoToNextRoomListener.asset
