+++ Assets/Resources/Singletons/GoToNextRoomListener.asset.meta
