+++ Assets/Resources/Singletons/PlayerInfo.asset.meta
