+++ Assets/Resources/Singletons/PlayerInfoStartGameListener.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: cf985c81376def94cb3a949dd2f1ab89
-folderAsset: yes
-DefaultImporter:
+guid: 3bcfb0a9dd77769469137ef6bbc8acf5
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
