+++ ProjectSettings/ProjectSettings.asset
@@ -122,6 +122,7 @@ PlayerSettings:
   preloadedAssets:
   - {fileID: 11400000, guid: 1eab53e6a0c54a0409f1fef37a0399a6, type: 2}
   - {fileID: 11400000, guid: 40e6a26e129b68447b24493f3788cd05, type: 2}
+  - {fileID: 11400000, guid: 2a6f2a7d74101b9419a81dd93d099aef, type: 2}
   metroInputSource: 0
   wsaTransparentSwapchain: 0
   m_HolographicPauseOnTrackingLoss: 1
